interface reorganization
