+++ CSRedis/CSRedis.csproj
@@ -91,8 +91,9 @@
     <Compile Include="Internal\Commands\RedisStatus.cs" />
     <Compile Include="Internal\Commands\RedisString.cs" />
     <Compile Include="Internal\Commands\RedisSubscription.cs" />
-    <Compile Include="IRedisClientAsync.cs" />
     <Compile Include="IRedisClient.cs" />
+    <Compile Include="IRedisClientAsync.cs" />
+    <Compile Include="IRedisClientSync.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="RedisClient.cs" />
     <Compile Include="Internal\RedisCommand.cs" />
