+++ CSRedis/RedisClient.cs
@@ -13,7 +13,7 @@ namespace CSRedis
     /// <summary>
     /// Represents a client connection to a Redis server instance
     /// </summary>
-    public partial class RedisClient : IRedisClient, IRedisClientAsync
+    public partial class RedisClient : IRedisClientSync, IRedisClientAsync
     {
         const int DefaultPort = 6379;
         const int DefaultConcurrency = 1000;
@@ -214,7 +214,7 @@ namespace CSRedis
         /// <typeparam name="T">Response type</typeparam>
         /// <param name="destination">Destination stream</param>
         /// <param name="func">Client command to execute (BULK reply only)</param>
-        public void StreamTo<T>(Stream destination, Func<IRedisClient, T> func)
+        public void StreamTo<T>(Stream destination, Func<IRedisClientSync, T> func)
         {
             StreamTo(destination, DefaultBufferSize, func);
         }
@@ -226,7 +226,7 @@ namespace CSRedis
         /// <param name="destination">Destination stream</param>
         /// <param name="bufferSize">Size of buffer used to write server response</param>
         /// <param name="func">Client command to execute (BULK reply only)</param>
-        public void StreamTo<T>(Stream destination, int bufferSize, Func<IRedisClient, T> func)
+        public void StreamTo<T>(Stream destination, int bufferSize, Func<IRedisClientSync, T> func)
         {
             _streaming = true;
             func(this);
